html: Revert "do not render line breaks between HTML tags"
This was not a good fix. There is plenty of HTML that would have been
parsed unproperly, e.g.

	This is a simple example of <em>HTML</em>
	that would have not been broken in two lines.

The proper fix is to always print line breaks instead of newlines, and
ensure that Sundown never inserts newlines when rendering span-level
elements, which is the case since the previous commit.