Started an extension to "outline" documents.
This is a rewrite of the outline extension for Python-Markdown
(http://git.constantvzw.org/?p=aa.mdx_outline.git;a=tree).

Wraps the document logical sections (as implied by h1-h6 headings) in
html5 <section> tags.

Big thank you to Pierre Marchand for helping me with this, and taking
time showing me the basics of C99!